refactored project structure (in Goland)
