+++ README.md
@@ -1,0 +1,18 @@
+# ece465-cloud-computing
+ECE465 Cloud Computing Homework Assignments
+
+Jonathan Lam & Henry Son
+
+Prof. Marano
+
+---
+
+### Build Instructions
+From the current directory, run:
+```bash
+GOPATH=$(pwd) go run src/projX/projX.go
+```
+where `X` is the project number. For more information about each project,
+see the README in `src/projX/` folder.
+
+TODO: elaborate here
