+++ src/proj1/README.md
@@ -1,0 +1,3 @@
+# ECE465 Project 1
+
+TODO: elaborate here
