+++ src/proj1/graph/graph.go
@@ -1,9 +1,9 @@
-package main
+package graph
 
 // Node represents a node of a Graph object
 type Node struct {
 	value int
-	edges []*Node
+	edges []int
 }
 
 // Graph represents a generic graph data structure
@@ -13,11 +13,11 @@ type Graph struct {
 
 // AddNode adds a node to a graph
 func (g *Graph) AddNode(value int) {
-	g.nodes = append(g, Node{value, make([]nodes, 0)})
+	g.nodes = append(g.nodes, Node{value, make([]int, 0)})
 }
 
 // AddUndirectedEdge adds an undirected edge between two nodes in a graph
-func (g *Graph) AddUndirectedEdge(n1, n2 *Node) {
-	g.nodes[n1] = append(g.nodes[n1], n2)
-	g.nodes[n2] = append(g.nodes[n2], n1)
+func (g *Graph) AddUndirectedEdge(n1, n2 int) {
+	g.nodes[n1].edges = append(g.nodes[n1].edges, n2)
+	g.nodes[n2].edges = append(g.nodes[n2].edges, n1)
 }
